refactor: split shared email layout
