+++ frontend/lib/emails/report-email-layout/email-doc-link.tsx
@@ -1,0 +1,10 @@
+import { Link } from "@react-email/components";
+import type { ReactNode } from "react";
+
+export function EmailDocLink({ href, children }: { href: string; children: ReactNode }) {
+  return (
+    <Link href={href} className="text-email-primary-300 no-underline">
+      {children}
+    </Link>
+  );
+}
