+++ frontend/lib/emails/report-email-layout/index.tsx
@@ -1,0 +1,6 @@
+export { EmailAction } from "./email-action";
+export { EmailDocLink } from "./email-doc-link";
+export { EmailFooter } from "./email-footer";
+export { EmailParagraph } from "./email-paragraph";
+export { EmailRows } from "./email-rows";
+export { LAMINAR_LOGO_CID, ReportEmailLayout } from "./report-email-layout";
