+++ frontend/public/laminar-logo-sm.png
